♻️ Refactor anonymization module: split pdf and docx internals by format
