+++ aymurai/text/anonymization/docx/__init__.py
@@ -1,0 +1,3 @@
+from aymurai.text.anonymization.docx.anonymizer import DocxAnonymizer
+
+__all__ = ["DocxAnonymizer"]
