+++ aymurai/text/anonymization/docx/anonymizer.py
@@ -1,9 +1,11 @@
 import os
 import tempfile
+from datetime import datetime, timezone
 from glob import glob
 from pathlib import Path
 from typing import Any
 
+from docx import Document
 from more_itertools import flatten
 
 from aymurai.text.anonymization.alignment import (
@@ -15,13 +17,29 @@ from aymurai.text.anonymization.base import (
     InvalidDocumentAnonymizer,
     register_anonymizer,
 )
-from aymurai.text.anonymization.watermarks import add_footer_watermark
-from aymurai.text.anonymization.xml_docx import (
+from aymurai.text.anonymization.docx.watermark import add_footer_watermark
+from aymurai.settings import settings
+from aymurai.text.anonymization.docx.xml import (
     create_docx,
     replace_text_in_xml,
     unzip_document,
 )
 from aymurai.utils.cache import cache_load, cache_save, get_cache_key
+
+
+def _set_aymurai_core_properties(doc_path: str) -> None:
+    """
+    Applies the configured AymurAI tooling metadata fields to the DOCX core properties.
+
+    Args:
+        doc_path (str): The path to the DOCX document to update.
+    """
+    document = Document(doc_path)
+    core_properties = document.core_properties
+    core_properties.author = ""
+    core_properties.last_modified_by = settings.ANONYMIZATION_METADATA_CREATOR
+    core_properties.modified = datetime.now(timezone.utc)
+    document.save(doc_path)
 
 
 @register_anonymizer
@@ -42,6 +60,19 @@ class DocxAnonymizer(BaseAnonymizer):
         output_dir: str = ".",
         render_context: dict[str, Any] | None = None,
     ) -> str:
+        """
+        Anonymizes a DOCX document using the matched paragraph predictions.
+
+        Args:
+        item (dict): The item dictionary containing the input DOCX path.
+        preds (list[dict]): The predictions to apply to the document.
+        output_dir (str, optional): The directory where the anonymized document should be written. Defaults to '.'.
+        render_context (dict[str, Any] | None, optional): The rendering context used to resolve replacement tokens.
+            Defaults to None.
+
+        Returns:
+            str: The path to the anonymized DOCX output file.
+        """
         item_path = Path(item["path"])
         file_path = self.ensure_file(item_path)
 
@@ -81,7 +112,8 @@ class DocxAnonymizer(BaseAnonymizer):
                 output_path = f"{output_dir}/{os.path.basename(str(file_path))}"
                 create_docx(tempdir, output_path)
 
-                # Add watermark to the footer
+                # Add metadata branding and the footer watermark
+                _set_aymurai_core_properties(output_path)
                 add_footer_watermark(output_path)
 
         if self.use_cache:
