+++ aymurai/text/anonymization/pdf/__init__.py
@@ -1,0 +1,3 @@
+from aymurai.text.anonymization.pdf.anonymizer import PdfAnonymizer
+
+__all__ = ["PdfAnonymizer"]
